+++ packages/react/src/components/DataTable/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 DataTable is a highly versatile component used to organize large quantities of tabular data. It supports pagination to help navigate through the data easily and efficiently, and also allows the user to sort and filter the data by columns. Additionally, it offers the ability to select one or multiple rows using checkboxes, which can be useful for performing batch actions or manipulating the data in other ways.
 
+## 2023-07-18 `1.2.3`
+
+#### 🐛 Bug fixes
+
+- Adjusted the `pagination` property to contain all properties available in the `DataTable.Footer` component. ([#69](https://github.com/TiendaNube/nimbus-patterns/pull/69) by [@juniorconquista](https://github.com/juniorconquista))
+
 ## 2023-06-15 `1.2.2`
 
 #### 💡 Others